// src/markdown/types.ts
export interface TableOfContents {
  tocFilePath: string;
  tocDirPath: string;
  rootPagesDir: string;
  root: TableOfContentsNode[];
}

export interface TableOfContentsNode {
  text?: string;
  url?: string;
  file?: string;
  hasParent?: boolean;
  children?: TableOfContentsNode[];
  depth: number;
}